- OXListBox: allow the specification of a background color other than the   default document background color.

+++ xclass/include/xclass/OXFrame.h
@@ -112,7 +112,7 @@ public:
 
   virtual int IsComposite() const { return False; }
 
-  virtual void SetBackgroundColor(unsigned int color)
+  virtual void SetBackgroundColor(unsigned long color)
        { OXWindow::SetBackgroundColor(_backPixel = color); }
 
   virtual int HandleEvent(XEvent *event);
